requirements_visualization.txt
# Requirements for Lunar Alighting RL Visualization Dashboard
# Install with: pip install -r requirements_visualization.txt

# Core data processing and visualization
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# JSON and data handling
ujson>=5.0.0  # Fast JSON parsing (optional but recommended)

# ZMQ for real-time data collection
pyzmq>=22.0.0

# Command line argument parsing
argparse  # Built-in with Python 3.7+

# Threading and timing (built-in modules)
# threading  # Built-in
# time       # Built-in
# datetime   # Built-in
# pathlib    # Built-in

# Optional: For better performance with large datasets
# scipy>=1.7.0
# scikit-learn>=1.0.0

# Optional: For interactive dashboards
# plotly>=5.0.0
# dash>=2.0.0
# streamlit>=1.0.0

# Optional: For advanced statistical analysis
# statsmodels>=0.13.0

# Optional: For better image quality in plots
# pillow>=8.0.0
